+++ purec/HHKCurrMechanism.cuh
@@ -37,10 +37,9 @@ extern __device__ mech_fun_t HHKCurrMechanism_mech_fxn_t;
 
 extern __device__ double HHKCurrMechanism_cuda_mech_fun(void* _self,
                                                         void* pre_comp,
-                                                        void* post_comp,
-                   
+                                                        void* post_comp,                   
                                                         const double global_time,
-                                                        const uint64_t curr_step);
+                                                        const uint_fast32_t curr_step);
 
 #endif /* CUDA */
 #endif /* HHKCURRMECHANISM_CUH */
