+++ purec/Mechanism.cuh
@@ -47,7 +47,7 @@ extern __device__ double cuda_mechanism_fxn(void* _self,
                                             void* pre_comp,
                                             void* post_comp,
                                             const double global_time,
-                                            const uint64_t curr_step);
+                                            const uint_fast32_t curr_step);
 
 #endif /* CUDA */
 #endif /* MECHANISM_CUH */
